backend/pyproject.toml
[project]
name = "finance-qa-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=0.2",
    "langchain-chroma>=0.1",
    "openai>=1.50",
    "yfinance>=0.2",
    "pandas>=2.2",
    "pandas-ta>=0.3",
    "redis>=5.0",
    "apscheduler>=3.10",
    "tavily-python>=0.5",
    "newsapi-python>=0.2",
    "rank-bm25>=0.2",
    "jieba>=0.42",
    "onnxruntime>=1.17",
    "transformers>=4.40",
    "numpy>=1.26",
    "pydantic-settings>=2.5",
    "aiosqlite>=0.20",
    "opentelemetry-proto>=1.35",
    "opentelemetry-exporter-otlp-proto-grpc>=1.35",
    "python-docx>=1.1",
    "pymupdf>=1.24",
    "wikipedia>=1.4.0",
    "optimum[onnxruntime]>=2.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.7",
    "mypy>=1.12",
    "optimum>=1.19",
    "torch>=2.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[[tool.uv.index]]
url="https://pypi.tuna.tsinghua.edu.cn/simple"
default=true